Fix t2i_adapter_canny_sdxl dtype mismatch: cast pipeline to float32
pipe.to("cpu", dtype=torch.float32) ensures all components (including
those with mixed precision from from_pretrained) are uniformly float32,
preventing c10::Half != float errors during encode_prompt.

+++ t2i_adapter_canny_sdxl/pytorch/src/model_utils.py
@@ -51,7 +51,7 @@ def load_t2i_adapter_canny_sdxl_pipe(adapter_model_name, base_model_name):
         torch_dtype=torch.float32,
     )
 
-    pipe.to("cpu")
+    pipe.to("cpu", dtype=torch.float32)
 
     modules = [
         pipe.text_encoder,
